+++ src/app/DisplayQuizzesUseCaseFactory.java
@@ -1,9 +1,9 @@
 package app;
 
 import interface_adapter.ViewManagerModel;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesController;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesPresenter;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesViewModel;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesController;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesPresenter;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesViewModel;
 import use_case.quiz.DisplayQuizzesInputBoundary;
 import use_case.quiz.DisplayQuizzesInteractor;
 import use_case.quiz.DisplayQuizzesOutputBoundary;
