+++ src/app/NoteUseCaseFactory.java
@@ -3,9 +3,9 @@ package app;
 import data_access.NoteDataAccessObject;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.note.*;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesController;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesPresenter;
-import interface_adapter.quiz.display_quiz.DisplayQuizzesViewModel;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesController;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesPresenter;
+import interface_adapter.quiz.display_quizzes.DisplayQuizzesViewModel;
 import use_case.note.NoteDataAccessInterface;
 import use_case.note.delete.DeleteNoteInputBoundary;
 import use_case.note.delete.DeleteNoteInteractor;
