+++ src/use_case/quiz/display_quizzes/DisplayQuizzesInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.quiz.display_quiz;
+package use_case.quiz.display_quizzes;
 
 public interface DisplayQuizzesInputBoundary {
     void execute();
