+++ src/use_case/quiz/display_quizzes/DisplayQuizzesInputData.java
@@ -1,4 +1,4 @@
-package use_case.quiz.display_quiz;
+package use_case.quiz.display_quizzes;
 
 public class DisplayQuizzesInputData {
     //TODO: this class is not needed, we can delete it later
